Fixed findMember.js and added a new search option.
I saw that $findMember gets the id of the replied member whom u replied keeping mention on so i have fixed that issue and also added the new option in which it will return a matching name with the member so like if you want to search a user named `Epic Games` you can just do `$findMember[Epic;no]` and it would return that member if there is none others with the same starting name. I hope it would be good for many. Also i tested it many times but didn't found any bugs but if u find the bug please inform me.